Dynamically tag Docker image with Gauge version
To get this working it was important to remove the unnecessary `-t`
tty flag from the `docker run` command.  When running with tty
then a carriage return is added at the end of every line. [1][2]
During the build pipeline we do not need an interactive terminal so we
should not be using the tty option anyway.

[1] https://stackoverflow.com/a/55109365
[2] https://github.com/moby/moby/issues/8513#issuecomment-216269864