Tag gauge properly (#30)
* Do not suppress errors in gauge version script

Also now using the shell helpers supplied by Mage [1]

And set the project up properly to use Go modules, by adding `go.mod`
and `go.sum` [2]

[1] https://github.com/magefile/mage/blob/master/sh/cmd.go
[2] https://github.com/golang/go/wiki/Modules#should-i-commit-my-gosum-file-as-well-as-my-gomod-file

* Dynamically tag Docker image with Gauge version

To get this working it was important to remove the unnecessary `-t`
tty flag from the `docker run` command.  When running with tty
then a carriage return is added at the end of every line. [1][2]
During the build pipeline we do not need an interactive terminal so we
should not be using the tty option anyway.

[1] https://stackoverflow.com/a/55109365
[2] https://github.com/moby/moby/issues/8513#issuecomment-216269864