Use Go for all Docker tags logic
Added end-to-end tests.

The tests stub out the call to the docker image.  One way to make it
even more of an end-to-end test would be to have a docker image
pre-installed on the base image just for the tests - as that would allow
us to make the actual calls to get the version numbers for the Docker
tags.