Run Go tests on CircleCI
Test results and code coverage are both available in the CircleCI UI,
as test results and build artifacts respectively.

The testing config can be extracted to a CircleCI orb and/or base image
at a later date.

The config is inspired by the config in Jenkins X:
https://github.com/jenkins-x/jx/blob/master/Makefile

It is important that the test script uses `set -o pipefail`, otherwise
the Circle build would pass even if tests failed.
See:
https://circleci.com/docs/2.0/using-shell-scripts/#set-error-flags

Command to run the tests locally:
`INTEGRATION_TEST=a CIRCLE_BUILD_NUM=1 go test -tags integration ./...`