Create a Go modules release on git tag push (#56)
Go modules dictates[1] that we should publish a new semantic version[2]
for every trunk commit (even if our module is not intended for use by
others).  This is a good thing.

Use GoReleaser[3] to create a new release every time a Git tag is pushed
to GitHub.

The current workflow is that after a pull request has been merged to
master, and the CircleCI build on master has passed, we then manually
create a new Git tag with our new semantic version, and push it to
GitHub.  This then automatically triggers the creation of the release.

An improvement on this workflow would be to indicate the release type
in the pull request (e.g. major, minor, patch), and then automatically
create the tag in the CircleCI build on master.

[1] https://blog.golang.org/publishing-go-modules
[2] https://semver.org/
[3] https://goreleaser.com/ci/