Ensure release job is visible in CircleCI UI
The release job looked to be triggering ok on git tag push, as it
should, but it was not displaying on the CircleCI UI.

Instead there was a build saying "No workflow", which could not be
navigated to.

Fixed by running the release job in the same workflow as the other
jobs.