Fix go lint problems (#74)
* Set ReviewDog reporting level to `error`

We want a reporting level of `error` so that we have to fix any reported
issues before pull requests can be merged.  This is fine as it is a
small greenfield codebase, which means we can follow the "no broken
windows"[1] principle. If we kept the reporting level as `warning` then
they could easily be ignored in the pull requests. We can tweak the
linting rules as we go to remove any that are too harsh.

 [1] https://www.freecodecamp.org/news/pragmatic-programmer-broken-windows-6916998eecbe/

* Fix go linting reported problems